refactoring and repair of a bug
