+++ clientSide/src/components/EditCustomer/EditCustomer.jsx
@@ -5,7 +5,6 @@ import GrayContainer from "../UiComponents/GrayContainer";
 import LabeledInput from "../LabeledInput/LabeledInput";
 import { onChangeIntHandler } from "../../functions/onChangeHandlers";
 import { useState } from "react";
-import { ColumnLabels } from "../../text-labels/parse-labels";
 import {
   CustomerInfoLabels,
   GeneralLabels,
@@ -13,42 +12,23 @@ import {
 import Button from "../Button/Button";
 import PageTitle from "../PageTitle/PageTitle";
 import { onChangeHandler } from "../../functions/onChangeHandlers";
-import Parse from "parse";
 import "./EditCustomer.css";
+import { updateCustomer } from "../../parse-functions/updateFunctions";
 
 const EditCustomer = () => {
   const { bookingId, customerId } = useParams();
   const navigate = useNavigate();
   const [formData, setFormData] = useState();
-  console.log(customerId);
-  const updateCustomer = async (e, customerId, formData, navigate) => {
-    e.preventDefault();
 
-    const Customer = Parse.Object.extend("Customer");
-    const query = new Parse.Query(Customer);
-    const object = await query.get(customerId);
-    object.set("lastName", formData.lastName);
-    object.set("driversLicenseID", formData.driversLicenseNo);
-    object.set("phoneNumber", formData.phoneNumber);
-    object.set("firstName", formData.firstName);
-    object.set("email", formData.email);
-    object.set("age", formData.age);
-    object.set("address", formData.address);
-    try {
-      const response = await object.save();
-      console.log("Customer updated", response);
-      navigate(`/individualBooking/${bookingId}`);
-    } catch (error) {
-      console.error("Error while updating Customer", error);
-    }
-  };
-  console.log(formData);
   return (
     <>
       <NavBar />
-
       <PageTitle ptitle="Edit customer"></PageTitle>
-      <form onSubmit={(e) => updateCustomer(e, customerId, formData, navigate)}>
+      <form
+        onSubmit={(e) =>
+          updateCustomer(e, customerId, bookingId, formData, navigate)
+        }
+      >
         <GrayContainer>
           <GrayColumn>
             <LabeledInput
